+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Structure generation module using packmol.
 - Provide explicit example of model comp ids in documentation.
 
 ## [0.0.4] - 2022-12-12